requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
chardet==5.2.0
langdetect==1.0.9
lxml==4.9.3
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.1.0